requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
fastapi==0.128.4
greenlet==3.3.1
h11==0.16.0
idna==3.11
jinja2==3.1.6
markupsafe==3.0.3
numpy==2.4.2
pandas==3.0.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-core==2.41.5
python-dateutil==2.9.0.post0
python-multipart==0.0.22
six==1.17.0
sqlalchemy==2.0.46
starlette==0.52.1
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.40.0
authlib==1.3.0
itsdangerous==2.1.2
httpx==0.27.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.156.0
